+++ R/learner_lightgbm_regr_lightgbm.R
@@ -178,6 +178,9 @@ LearnerRegrLightGBM = R6Class("LearnerRegrLightGBM",
             if (is.null(param_vals$early_stopping_rounds)) {
               stop("Parameter 'early_stopping_rounds' must be set to use internal tuning.")
             }
+            if (is.null(param_vals$eval)) {
+              stop("Parameter 'eval' must be explicitly set for internal tuning.")
+            }
             assert_integerish(domain$upper, len = 1L, any.missing = FALSE)
           }, .parent = topenv()),
           disable_in_tune = list(