+++ R/learner_xgboost_surv_xgboost_aft.R
@@ -6,7 +6,7 @@
 #' eXtreme Gradient Boosting regression using an **Accelerated Failure Time**
 #' objective.
 #' Calls [xgboost::xgb.train()] from package \CRANpkg{xgboost} with `objective`
-#' set to `survival:aft` and `eval_metric` to `aft-nloglik` by default.
+#' set to `survival:aft` and `eval_metric` to `aft-nloglik`.
 #'
 #' @details
 #' This learner returns three prediction types:
@@ -19,7 +19,7 @@
 #' @template note_xgboost
 #'
 #' @section Initial parameter values:
-#' - `nrounds` is initialized to 1.
+#' - `nrounds` is initialized to 1000.
 #' - `nthread` is initialized to 1 to avoid conflicts with parallelization via \CRANpkg{future}.
 #' - `verbose` is initialized to 0.
 #'
@@ -124,7 +124,7 @@ LearnerSurvXgboostAFT = R6Class("LearnerSurvXgboostAFT",
       ps$add_dep("top_k", "feature_selector", CondAnyOf$new(c("greedy", "thrifty")))
 
       # custom defaults
-      ps$values = list(nrounds = 1L, nthread = 1L, verbose = 0L)
+      ps$values = list(nrounds = 1000L, nthread = 1L, verbose = 0L)
 
       super$initialize(
         id = "surv.xgboost.aft",