+++ NEWS.md
@@ -8,11 +8,12 @@
   the upstream package again.
 * Add early stopping and validation support to learners `lrn("surv.xgboost.aft")`
   and `lrn("surv.xgboost.cox")`.
-* Added early stopping and validation to catboost and lightgbm. 
+* Added early stopping and validation to catboost and lightgbm.
 * Added missing `case.depth` parameter to `rfsrc` learners.
 * `mlr3` is now in Depends instead of Imports.
 * Deprecated learner `lrn("surv.xgboost")` was now removed.
   Use `lrn("surv.xgboost.cox")` or `lrn("surv.xgboost.aft")` instead.
+* Change xgboost default nrounds from 1 to 1000.
 
 # mlr3extralearners 0.8.0